added safety features

+++ main.py
@@ -1,69 +1,81 @@
+# system modules
+import numpy as np
+import json
+import os.path
+# custom modules
 import bd
 import menu
-import numpy as np
-import pprint
-import json
 
+# Global variables
 # array of main menu items
 main_m = np.array(["Configure beam", "Configure loads",
                    "Save beam and loads", "Load beam and loads", "Generate plot", "Quit"])
 # variables storing beam settings
 bl = 0.
 bs = "both"
-# create empty array with zero rows and two columns
+# empty array with zero rows and two columns used for storing the loads and its positions
 l = np.empty(shape=[0, 2], dtype=float)
 
 
 ##
-# @brief      Asks user to input new load values
+## @brief      Asks user to input new load values
 ##
-# @param      loadArray  the array that should be appended with new load values
+## @param      loadArray  the array that should be appended with new load values
 ##
-# @return     returns the array with new load input
+## @return     the array with new load input
 ##
-# @author     Matej Majtan
+## @author     Matej Majtan
 ##
 def addLoad(loadArray):
     lp = 0
     lf = menu.inputNumber("Input force of the load: ", "f")
     while lp <= 0 or lp > bl:
         lp = menu.inputNumber("Input position of the load: ", "f")
+        if lp <= 0 or lp >bl:
+        	print("Please enter a position that is in the range of the beam lenght")
     print()     # new line
     return np.append(loadArray, [[lf, lp]], axis=0)
 
 
 ##
-# @brief      Saves a file.
+## @brief      Saves a file.
 ##
-# @param      filename     Name of the file data should be saved into (.json at the end is optional)
-# @param      beamLenght   The beam lenght
-# @param      beamSupport  The beam support
-# @param      loadArray    The load array
+## @param      filename     Name of the file data should be saved into (.json at the end is optional)
+## @param      beamLenght   the length of the beam that should be saved
+## @param      beamSupport  the type of the beam support that should be saved
+## @param      loadArray    the numpy array of loads and its positions that should be saved
 ##
-# @author     Matej Majtan
+## @author     Matej Majtan
 ##
 def saveFile(filename, beamLenght, beamSupport, loadArray):
+	# reate a dictionary used by the json module to create the json file
     data = {"beamLenght": beamLenght,
             "beamSupport": beamSupport,
             "loadArray": [list(i) for i in loadArray]
             }
 
+    # open a file and dump the json data into it
     with open(filename if filename.endswith(".json") else filename + ".json", "w") as write_file:
         json.dump(data, write_file)
 
 
 ##
-# @brief      Loads a file.
+## @brief      Loads a json file  to the system
 ##
-# @param      filename  Name of the file data should be loaded from (.json at the end is optional)
+## @param      filename  Name of the file data should be loaded from (.json at the end is optional)
 ##
-# @return     tuple with leaded data for beam length, beam support and load array
+## @return     tuple with leaded data for beam length, beam support and load array
+##
+## @author     Matej Majtan
 ##
 def loadFile(filename):
     data = {}
     with open(filename if filename.endswith(".json") else filename + ".json", "r") as read_file:
         data = json.load(read_file)
+
+    # TODO: add checking for correct file (data in it)
     return (data["beamLenght"], data["beamSupport"], np.array([np.array(i) for i in data["loadArray"]]))
+
 
 # main loop start
 while True:
@@ -75,9 +87,11 @@ while True:
         bl = 0
         while bl <= 0:
             bl = menu.inputNumber("Please enter a valid beam length: ", "f")
+            if bl <= 0:
+            	print("The beam lenght has to be a positive number!")
 
         # ask user which beam support should be used
-        print("Choose the beam support:\n")
+        print("\nChoose the beam support:")
         bs_m = menu.displayMenu(["both", "cantilever"])
         if bs_m == 1:
             bs = "both"
@@ -91,8 +105,8 @@ while True:
             if np.any(l):
                 # show the current loads
                 print("Current configured loads:")
-                print("force\tposition")
-                print("".join(["{}\t{}\n".format(i, j) for i, j in l]))
+                print("force\t\tposition")
+                print("".join(["{}\t\t{}\n".format(i, j) for i, j in l]))
 
                 # give an option for adding or removing
                 lm = menu.displayMenu(["Add a load", "Remove a load", "Back"])
@@ -103,9 +117,15 @@ while True:
                 # remove load chosen
                 elif lm == 2:
                     print("Please choose entry you wish to delete:")
-                    rm = menu.displayMenu(
-                        ["{}\t{}".format(i, j) for i, j in l])
-                    l = np.delete(l, rm - 1, 0)
+                    print("   force\tposition")
+                    # create a list of the current loads and their positions
+                    rm_menu = ["{}\t{}".format(i, j) for i, j in l]
+                    rm_menu.append("Back")
+                    # show the menu
+                    rm = menu.displayMenu(rm_menu)
+                    # check whether the back option hasn't been chosen
+                    if rm != len(rm_menu):
+                    	l = np.delete(l, rm - 1, 0)
                 # back chosen
                 elif lm == 3:
                     break
@@ -116,6 +136,7 @@ while True:
 
                 # add load chosen
                 if lm == 1:
+                	# check whether the beam length was given
                     if bl == 0:
                         print("You need to first specify the beam length")
                     else:
@@ -126,15 +147,18 @@ while True:
 
     # Save beam and loads option chosen
     elif opt == 3:
-        file = input(
-            "Please enter the file name (.json will be added automatically): ")
+        file = input("Please enter the file name (*.json sufix will be added automatically): ")
         saveFile(file, bl, bs, l)
         print("Data successfully saved")
     # Load beam and loads option chosen
     elif opt == 4:
         file = input("Please enter the file name: ")
-        bl, bs, l = loadFile(file)
-        print("Data successfully loaded")
+        # check whether the file exists
+        if os.path.isfile(file if file.endswith(".json") else file + ".json")
+        	bl, bs, l = loadFile(file)
+        	print("Data successfully loaded")
+        else:
+        	print("No such file found")
     # Generate plot option chosen
     elif opt == 5:
         # plot the given values
